Simplify MessageSource based on user feedback
The MessageSource interface design made is hard to deal with a variety
of error cases elegantly. ConsumeMessages was blocking, and to exit the
consumption loop, the caller had to call Close() concurrently.
It was never clear whether the lifecycle of the consumption process
lived with with the ConsumeMessages call, or the MessageSource instance
this was confusing.  It also meant that the caller had to know that the
MessageSource was effectively closed when exiting on error, even though
Close() was never called explicitly.

This has been simplified to use the new standard context.Context
mechanism for exiting ConsumeMessages().  The Close() method has gone
from MessageSource.

Unfortunately, this is an API change, but can simplify message
consumption code significantly.