Replace "broken" int32 with a channel
This replaces the broken variable with a channel instead of an int32 and
sync.Atomic for accessing it.  This change alone doesn't fix or change
much, but it's a move towards moving more of the synchronisation to
using channels in order to simplify the code and move to a single step
synchronisation process within the callback.