Replace "active" WaitGroup with a channel
This replaces the "active" WaitGroup with a channel, meaning that all
synchronisation is now done via channels, and that there is a single
select block in the callback rather than a two step synchronisation
process.

Note that this differs slightly in design in that only one callback can
be active at once, where the old design allowed many.  This is fine
though, since the nats-streaming API doesn't call this function multiple
times concurrently anyway.  (If we did get called multiple times
concurrently, we'd also be calling back the consumer application
callback multiple times concurrently, and what does that even mean in an
API that guarantees ordering?)